refactor: split firebase helpers into separate files
